framework: For the sysdetect component use --enable and --disable at configure and do not show native event output for utils
