+++ src/components/sysdetect/sysdetect.c
@@ -469,7 +469,7 @@ papi_vector_t _sysdetect_vector = {
     .cmp_info = {
                  .name = "sysdetect",
                  .short_name = "sysdetect",
-                 .description = "System info detection component",
+                 .description = "Query system information through papi_hardware_avail or using the PAPI API (see the source files under sysdetect/tests)",
                  .version = "1.0",
                  .support_version = "n/a",
                  .kernel_version = "n/a",
