+++ src/utils/Makefile
@@ -16,8 +16,8 @@ ALL = papi_avail papi_mem_info papi_cost papi_clockres papi_native_avail \
 
 default all utils: $(ALL)
 
-papi_avail: papi_avail.o $(PAPILIB) print_header.o
-	$(CC) $(CFLAGS) $(OPTFLAGS) -o papi_avail papi_avail.o print_header.o $(PAPILIB) $(LDFLAGS)
+papi_avail: papi_avail.o $(PAPILIB) utils_helper.o
+	$(CC) $(CFLAGS) $(OPTFLAGS) -o papi_avail papi_avail.o utils_helper.o $(PAPILIB) $(LDFLAGS)
 
 papi_clockres: papi_clockres.o $(PAPILIB) $(CLOCKCORE)
 	$(CC) $(CFLAGS) $(OPTFLAGS) -o papi_clockres papi_clockres.o $(PAPILIB) $(CLOCKCORE) -lm $(LDFLAGS)
@@ -25,8 +25,8 @@ papi_clockres: papi_clockres.o $(PAPILIB) $(CLOCKCORE)
 papi_command_line: papi_command_line.o $(PAPILIB) $(DOLOOPS)
 	$(CC) $(CFLAGS) $(OPTFLAGS) -o papi_command_line papi_command_line.o $(PAPILIB) $(DOLOOPS) $(LDFLAGS)
 
-papi_component_avail: papi_component_avail.o $(PAPILIB) print_header.o
-	$(CC) $(CFLAGS) $(OPTFLAGS) -o papi_component_avail papi_component_avail.o $(PAPILIB) print_header.o $(LDFLAGS) $(LIBSDEFLAGS)
+papi_component_avail: papi_component_avail.o $(PAPILIB) utils_helper.o
+	$(CC) $(CFLAGS) $(OPTFLAGS) -o papi_component_avail papi_component_avail.o $(PAPILIB) utils_helper.o $(LDFLAGS) $(LIBSDEFLAGS)
 
 papi_cost: papi_cost.o $(PAPILIB) cost_utils.o
 	$(CC) $(CFLAGS) $(OPTFLAGS) -o papi_cost papi_cost.o cost_utils.o $(PAPILIB) -lm $(LDFLAGS)
@@ -37,8 +37,8 @@ papi_decode: papi_decode.o $(PAPILIB)
 papi_error_codes: papi_error_codes.o $(PAPILIB)
 	$(CC) $(CFLAGS) $(OPTFLAGS) -o papi_error_codes papi_error_codes.o $(PAPILIB) $(LDFLAGS)
 
-papi_event_chooser: papi_event_chooser.o $(PAPILIB) print_header.o
-	$(CC) $(CFLAGS) $(OPTFLAGS) -o papi_event_chooser papi_event_chooser.o print_header.o $(PAPILIB) $(LDFLAGS)
+papi_event_chooser: papi_event_chooser.o $(PAPILIB) utils_helper.o
+	$(CC) $(CFLAGS) $(OPTFLAGS) -o papi_event_chooser papi_event_chooser.o utils_helper.o $(PAPILIB) $(LDFLAGS)
 
 papi_hybrid_native_avail: papi_hybrid_native_avail.o $(PAPILIB)
 	$(CC) $(CFLAGS) $(OPTFLAGS) -o papi_hybrid_native_avail papi_hybrid_native_avail.o $(PAPILIB) $(LDFLAGS)
@@ -49,11 +49,11 @@ papi_mem_info: papi_mem_info.o $(PAPILIB)
 papi_multiplex_cost: papi_multiplex_cost.o $(PAPILIB) cost_utils.o
 	$(CC) $(CFLAGS) $(OPTFLAGS) -o papi_multiplex_cost papi_multiplex_cost.o cost_utils.o $(PAPILIB) -lm $(LDFLAGS)
 
-papi_hardware_avail: papi_hardware_avail.o $(PAPILIB) print_header.o
-	$(CC) $(CFLAGS) $(OPTFLAGS) -o papi_hardware_avail papi_hardware_avail.o $(PAPILIB) print_header.o $(LDFLAGS)
+papi_hardware_avail: papi_hardware_avail.o $(PAPILIB) utils_helper.o
+	$(CC) $(CFLAGS) $(OPTFLAGS) -o papi_hardware_avail papi_hardware_avail.o $(PAPILIB) utils_helper.o $(LDFLAGS)
 
-papi_native_avail: papi_native_avail.c $(PAPILIB) print_header.o
-	$(CC) $(CFLAGS) $(OPTFLAGS) $(INCLUDE) -o papi_native_avail papi_native_avail.c $(PAPILIB) print_header.o $(LDFLAGS) $(LIBSDEFLAGS)
+papi_native_avail: papi_native_avail.c $(PAPILIB) utils_helper.o
+	$(CC) $(CFLAGS) $(OPTFLAGS) $(INCLUDE) -o papi_native_avail papi_native_avail.c $(PAPILIB) utils_helper.o $(LDFLAGS) $(LIBSDEFLAGS)
 
 papi_version: papi_version.o $(PAPILIB)
 	$(CC) $(CFLAGS) $(OPTFLAGS) -o papi_version papi_version.o $(PAPILIB) $(LDFLAGS)
@@ -66,8 +66,8 @@ papi_xml_event_info: papi_xml_event_info.o $(PAPILIB)
 cost_utils.o: ../testlib/papi_test.h cost_utils.c
 	$(CC) $(INCLUDE) $(CFLAGS) $(OPTFLAGS) -c cost_utils.c
 
-print_header.o: print_header.h print_header.c
-	$(CC) $(INCLUDE) $(CFLAGS) $(OPTFLAGS) -c print_header.c
+utils_helper.o: utils_helper.h utils_helper.c
+	$(CC) $(INCLUDE) $(CFLAGS) $(OPTFLAGS) -c utils_helper.c
 
 clean:
 	rm -f *.o *.stderr *.stdout core *~ $(ALL)
