+++ src/utils/papi_hardware_avail.c
@@ -26,7 +26,7 @@
 #include <string.h>
 
 #include "papi.h"
-#include "print_header.h"
+#include "utils_helper.h"
 
 typedef struct command_flags
 {
