--- src/utils/print_header.h
@@ -1,2 +1,0 @@
-int papi_print_header( char *prompt, const PAPI_hw_info_t ** hwinfo );
-
