+++ src/utils/utils_helper.h
@@ -1,0 +1,7 @@
+#ifndef UTILS_HELPER_H
+#define UTILS_HELPER_H
+
+int papi_print_header(char *prompt, const PAPI_hw_info_t ** hwinfo);
+int check_for_utility_components(const char *component_name);
+
+#endif /* UTILS_HELPER_H */
